Create common layout to show values with border in details page Create layout-land folder to create landscape view screens
