+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/fragments/home/HomeFragment.kt
@@ -8,6 +8,7 @@ import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
@@ -19,9 +20,11 @@ import jp.co.yumemi.android.code_check.constants.DialogConstants
 import jp.co.yumemi.android.code_check.constants.MessageConstants
 import jp.co.yumemi.android.code_check.constants.MessageConstants.PROGRESS_DIALOG_MESSAGE
 import jp.co.yumemi.android.code_check.databinding.FragmentHomeBinding
+import jp.co.yumemi.android.code_check.interfaces.ConfirmDialogButtonClickListener
 import jp.co.yumemi.android.code_check.interfaces.CustomAlertDialogListener
 import jp.co.yumemi.android.code_check.models.GitHubRepoObject
 import jp.co.yumemi.android.code_check.utils.DialogUtils.Companion.showAlertDialog
+import jp.co.yumemi.android.code_check.utils.DialogUtils.Companion.showConfirmAlertDialog
 import jp.co.yumemi.android.code_check.utils.DialogUtils.Companion.showErrorDialogInFragment
 import jp.co.yumemi.android.code_check.utils.DialogUtils.Companion.showProgressDialogInFragment
 
@@ -53,8 +56,31 @@ class HomeFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        initView()
         initiateAdapter()
         viewModelObservers()
+    }
+
+    private fun initView() {
+        val callback = object : OnBackPressedCallback(true) {
+            override fun handleOnBackPressed() {
+                // Handle back button press for Home Fragment
+                showConfirmAlertDialog(
+                    requireActivity(),
+                    MessageConstants.EXIT_CONFIRMATION_MESSAGE,
+                    object : ConfirmDialogButtonClickListener {
+                        override fun onPositiveButtonClick() {
+                            requireActivity().finish()
+                        }
+
+                        override fun onNegativeButtonClick() {
+
+                        }
+                    })
+            }
+        }
+
+        requireActivity().onBackPressedDispatcher.addCallback(viewLifecycleOwner, callback)
     }
 
     /**
@@ -121,5 +147,10 @@ class HomeFragment : Fragment() {
                 gitHubRepo
             )
         )
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        binding = null
     }
 }
