+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/fragments/repodetails/RepoDetailsFragment.kt
@@ -12,6 +12,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.navArgs
 import jp.co.yumemi.android.code_check.databinding.FragmentRepoDetailsBinding
 import jp.co.yumemi.android.code_check.models.GitHubRepoObject
+import jp.co.yumemi.android.code_check.utils.UIUtils.Companion.changeUiSize
 
 /**
  * Details Page Fragment
@@ -20,7 +21,7 @@ import jp.co.yumemi.android.code_check.models.GitHubRepoObject
 class RepoDetailsFragment : Fragment() {
 
     private val args: RepoDetailsFragmentArgs by navArgs()
-    private lateinit var binding: FragmentRepoDetailsBinding
+    private var binding: FragmentRepoDetailsBinding? = null
     private lateinit var viewModel: RepoDetailsViewModel
     private lateinit var gitHubRepo: GitHubRepoObject
 
@@ -38,11 +39,12 @@ class RepoDetailsFragment : Fragment() {
          */
         binding = FragmentRepoDetailsBinding.inflate(inflater, container, false)
         viewModel = ViewModelProvider(requireActivity())[RepoDetailsViewModel::class.java]
-        binding.vm = viewModel
-        binding.lifecycleOwner = this
+        binding?.vm = viewModel
+        binding?.lifecycleOwner = this
 
-        return binding.root
+        return binding?.root
     }
+
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         setData()
@@ -51,12 +53,14 @@ class RepoDetailsFragment : Fragment() {
     /**
      * Pass selected Git Repo Object to view model
      * Set data to view
-     * @see https://bumptech.github.io/glide/
      */
     private fun setData() {
         viewModel.setGitRepoData(gitHubRepo)
 
-        /* Show profile icon using Glide */
-//        Glide.with(this).load(gitHubRepo.owner?.avatarUrl).into(binding.ownerIconView)
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        binding = null
     }
 }
