+++ app/src/main/res/navigation/nav_graph.xml
@@ -3,20 +3,20 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/apk/res-auto"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/oneFragment">
+    app:startDestination="@id/homeFragment">
 
     <fragment
-        android:id="@+id/oneFragment"
+        android:id="@+id/homeFragment"
         android:name="jp.co.yumemi.android.code_check.ui.fragments.home.HomeFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_home">
         <action
             android:id="@+id/action_repositoriesFragment_to_repositoryFragment"
-            app:destination="@id/twoFragment" />
+            app:destination="@id/repoDetailsFragment" />
     </fragment>
 
     <fragment
-        android:id="@+id/twoFragment"
+        android:id="@+id/repoDetailsFragment"
         android:name="jp.co.yumemi.android.code_check.ui.fragments.repodetails.RepoDetailsFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_repo_details">
