+++ app/src/main/res/values/dimens.xml
@@ -4,6 +4,7 @@
     <dimen name="vertical_margin">16dp</dimen>
     <dimen name="mini_horizontal_margin">8dp</dimen>
     <dimen name="mini_vertical_margin">8dp</dimen>
+    <dimen name="content_margin_top">8dp</dimen>
 
     /* Dialog View Dimensions */
     <dimen name="dialog_horizontal_padding">16dp</dimen>
@@ -22,4 +23,5 @@
     <dimen name="content_card_view_corner_radius">20dp</dimen>
     <dimen name="content_card_view_elevation">10dp</dimen>
     <dimen name="list_card_view_corner_radius">12dp</dimen>
+    <dimen name="avatar_margin">16dp</dimen>
 </resources>
