+++ app/src/main/res/values/themes.xml
@@ -27,6 +27,24 @@
         <item name="android:textColor">@color/white</item>
     </style>
 
+    <style name="nameTextViewStyle">
+        <item name="android:textSize">18sp</item>
+        <item name="android:textColor">@color/black</item>
+        <item name="android:textStyle">bold</item>
+    </style>
+
+    <style name="languageTextViewStyle">
+        <item name="android:textSize">14sp</item>
+        <item name="android:textColor">@color/black</item>
+    </style>
+
+    <style name="contentStyle">
+        <item name="android:textSize">20sp</item>
+        <item name="android:textColor">@color/black</item>
+        <item name="android:textStyle">bold</item>
+    </style>
+
+
     <!-- Progress Dialog Style-->
     <style name="progressBarStyle" parent="@style/Widget.AppCompat.ProgressBar.Horizontal">
         <item name="colorAccent">@color/accent_color</item>
